A few improvements
- Fix ahk script usage
- Refactor default files setup
- Add default input.json and registry.json
- Probably some other small things

+++ src/impyrium/helpers.py
@@ -1,12 +1,18 @@
 from .default_files import defaults
+from .default_files.ahk_get_file_script import ahkGetFile
+
+import os
 
 def getCurrentlySelectedFile():
     TEMP_FILE = "__currently_selected__.txt"
-    os.system(f"\"{defaults.AHK}\" {defaults.WINDOWS_SCRIPTS_FOLDER}/{defaults.GET_FILE_AHK_SCRIPT_FILE} {TEMP_FILE}")
 
-    f = open(TEMP_FILE)
-    result = f.read()
-    f.close()
+    if ahkGetFile.exists():
+        os.system(f"\"{defaults.AHK}\" {ahkGetFile.getPath()} {TEMP_FILE}")
 
-    os.remove(TEMP_FILE)
-    return result
+        f = open(TEMP_FILE)
+        result = f.read()
+        f.close()
+
+        os.remove(TEMP_FILE)
+        return result
+    return ""
